refactor: deploy and interactions in contracts-ts
