+++ packages/contracts-ts/src/deploy.ts
@@ -1,4 +1,4 @@
-import { Abi } from '@cowprotocol/common'
+import { Abi, AbstractProviderAdapter, AdapterTypes } from '@cowprotocol/common'
 /**
  * The contract used to deploy contracts deterministically with CREATE2.
  * The address is chosen by the hardhat-deploy library.
@@ -53,3 +53,32 @@ export interface NamedArtifactDeployment<C extends ContractName> extends Artifac
 export type MaybeNamedArtifactArtifactDeployment<C> = C extends ContractName
   ? NamedArtifactDeployment<C>
   : ArtifactDeployment
+
+export class ContractsTs_Deploy<T extends AdapterTypes = AdapterTypes> {
+  SALT: string
+  DEPLOYER_CONTRACT: string
+
+  constructor(private adapter: AbstractProviderAdapter<T>) {
+    this.adapter = adapter
+    this.SALT = this.SALT = this.adapter.formatBytes32String('Mattresses in Berlin!')
+    this.DEPLOYER_CONTRACT = '0x4e59b44847b379578588920ca78fbf26c0b4956c'
+  }
+
+  /**
+   * Computes the deterministic address at which the contract will be deployed.
+   * This address does not depend on which network the contract is deployed to.
+   *
+   * @param contractName Name of the contract for which to find the address.
+   * @param deploymentArguments Extra arguments that are necessary to deploy.
+   * @returns The address that is expected to store the deployed code.
+   */
+  public deterministicDeploymentAddress<C>(
+    adapter: AbstractProviderAdapter,
+    { abi, bytecode }: MaybeNamedArtifactArtifactDeployment<C>,
+    deploymentArguments: DeploymentArguments<C>,
+  ): string {
+    const deployData = adapter.hexConcat([bytecode, this.adapter.encodeDeploy(deploymentArguments, abi as T['Abi'])])
+
+    return adapter.getCreate2Address(DEPLOYER_CONTRACT, this.SALT, this.adapter.keccak256(deployData))
+  }
+}
